requirements.txt
tensorflow
matplotlib
pillow
faiss-cpu
supabase
flask[async]
python-dotenv
flask_cors
torch
torchvision
sentence_transformers
tf_keras
rembg
opencv-python